+++ gc3libs/template.py
@@ -1,5 +1,5 @@
 #! /usr/bin/env python
-#
+
 """
 Support and expansion of programmatic templates.
 
@@ -11,6 +11,7 @@ with a list of substitutions (using the syntax of Python's standard
 texts coming from the same template.  Templates can be nested, and
 expansions generated recursviely.
 """
+
 # Copyright (C) 2009-2012, 2014, 2019  University of Zurich. All rights reserved.
 #
 # This program is free software; you can redistribute it and/or modify
@@ -25,7 +26,7 @@ expansions generated recursviely.
 #
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
+
 from __future__ import absolute_import, print_function, unicode_literals
 from builtins import str
 from builtins import range
@@ -42,69 +43,7 @@ try:
 except:
     # use our own implementation, in case `itertools` does not (yet)
     # provide the set-product
-    class SetProductIterator(object):
-
-        """Iterate over all elements in a cartesian product.
-
-        Argument `factors` is a sequence, all whose items are sequences
-        themselves: the returned iterator will return -upon each
-        successive invocation- a list `[t_1, t_2, ..., t_n]` where `t_k`
-        is an item in the `k`-th sequence.
-
-        Examples::
-          >>> list(SetProductIterator([]))
-          [[]]
-          >>> list(SetProductIterator([1]))
-          [[1]]
-          >>> list(SetProductIterator([1],[1]))
-          [[1, 1]]
-          >>> list(SetProductIterator([1,2],[]))
-          [[]]
-          >>> list(SetProductIterator([1,2],[1]))
-          [[1, 1], [2, 1]]
-          >>> list(SetProductIterator([1, 2], [1, 2]))
-          [[1, 1], [2, 1], [1, 2], [2, 2]]
-        """
-
-        def __init__(self, *factors):
-            self.__closed = False
-            self.__factors = factors
-            self.__L = len(factors)
-            self.__M = [len(s) - 1 for s in factors]
-            self.__m = [0] * self.__L
-            self.__i = 0
-
-        def __iter__(self):
-            return self
-
-        def __next__(self):
-            if self.__closed:
-                raise StopIteration
-            if (0 == self.__L) or (-1 in self.__M):
-                # there are no factors, or one of them has no elements
-                self.__closed = True
-                return []
-            else:
-                if self.__i < self.__L:
-                    # will return element corresponding to current multi-index
-                    result = [s[self.__m[i]]
-                              for (i, s) in enumerate(self.__factors)]
-                    # advance multi-index
-                    i = 0
-                    while (i < self.__L):
-                        if self.__m[i] == self.__M[i]:
-                            self.__m[i] = 0
-                            i += 1
-                        else:
-                            self.__m[i] += 1
-                            break
-                    self.__i = i
-                    # back to caller
-                    return result
-                else:
-                    # at end of iteration
-                    self.__closed = True
-                    raise StopIteration
+    from gc3libs.compat.set_product_iterator import SetProductIterator
 
 
 class Template(object):
